# Faster __getattr__ in nn.Module