--- a/torch/nn/modules/module.py
+++ b/torch/nn/modules/module.py
@@ -1694,20 +1694,25 @@ class Module:
     # https://github.com/microsoft/pyright/issues/4213
     def __getattr__(self, name: str) -> Any:
         __dict__ = self.__dict__
-        _parameters = __dict__['_parameters']
+        _parameters = __dict__.get('_parameters')
         if _parameters:
-            result = _parameters.get(name, None)
-            if result is not None:
+            # A param can be None so we use False instead to check if the key
+            # is in the _parameters dict once and only once.
+            # We use False but any non-None, non-Parameter value would do.
+            # The alternative `if name in _parameters: return _parameters[name]`
+            # accesses the value of `name` twice when only one is required
+            result = _parameters.get(name, False)
+            if result is not False:
                 return result
-        _buffers = __dict__['_buffers']
+        _buffers = __dict__.get('_buffers')
         if _buffers:
-            result = _buffers.get(name, None)
-            if result is not None:
+            result = _buffers.get(name, False)
+            if result is not False:
                 return result
-        _modules = __dict__['_modules']
+        _modules = __dict__.get('_modules')
         if _modules:
-            result = _modules.get(name, None)
-            if result is not None:
+            result = _modules.get(name, False)
+            if result is not False:
                 return result
         raise AttributeError(f"'{type(self).__name__}' object has no attribute '{name}'")