Similar improvements could be done for __setattr__

```python
from torch import nn
import torch.utils.benchmark
import torch

class MyModule(nn.Module):
    def __init__(self, ):
        super().__init__()
        self.net = nn.ModuleList([])

mod = MyModule()

cmd = "mod.net"
print(torch.utils.benchmark.Timer(cmd, globals=globals()).adaptive_autorange())

```

Runs in 370ns with main, 264 ns with this PR
If the module has params:
292.25 ns (here) vs 360.81 ns (main)

cc @albanD @mruberry @jbschlosser @walterddr @mikaylagawarecki