@cyyever 2024-05-22
What if "_parameters__" not in __dict__?