@vmoens 2024-05-22
I thought about that but it seems `__init__` always creates it no?
If not I had another safer version with roughly similar speed ups.