@cyyever 2024-05-22
May be it is deleted later by existing code? Use the safer version, don't break the semantic.