@vmoens 2024-05-22
I made it more conservative for `__getattr__`
I also updated `__setattr__`. Previously, we checked if `"_parameters"` or `"_modules"` were not defined and if so we explicitly assume that this is because `super().__init__` has not been called.
Now the logic is as follows: first get `"_parameters"` or `None` if it isn't there. In the latter case, check if the value is a parameter or a module. If it is, raise an error (because `__init__` has not been called). If not just set that value in `__dict__` using super.
In the former case (the module is init), check if the value is a param, then check if the name is in param. Then do the same with module. Then with buffer.

Beyond trivially reducing the calls to `self.__dict__` we also do fewer checks and fewer repetitions of checks that are equivalent (like looking at whether params is None and then looking whether modules is None).

I forgot to say in the description that the goal of this PR is to reduce the overhead of functional calls where there are multiple getattr and setattr.